Hot patch 1.14 proxy URLs
This is really more of a design failure with how this is written versus
the concepts of the downloads API.

I do not have time to fix this _properly_ right now, however we
obviously want these downloads to be working and available for people.
So this ugly, nasty, really unacceptable change will do for now.