Added pronouns to ZTIs
A few unfortunately somewhat common occurrences in the community recently have included some people questioning the use of pronouns in another user's username, ridiculing the concept of putting pronouns in one's name, or putting fake pronouns in one's name (i.e. `(that/it)`). In some cases, the fake pronouns are meant sarcastically or as a joke, but it does not feel like a joke to the people being ridiculed for it. To help combat this, I am proposing adding a section to the CG regarding respecting people's desired pronouns. 

This is not: a blanket statement requiring everybody to be hyperaware of desired pronouns
This is: a statement requiring clearly visible pronouns to be respected