add player safety bypass rule
As agreed upon by the vast majority of the Paper team, this rule is added to forbid discussion of how to bypass player safety features such as chat reporting. As always, actually doing it will still be possible with Paper, but like with offline mode, it won't be supported or discussed in the community.